requirements.txt
pandas>=2.1,<3
numpy>=1.26,<2
scikit-learn>=1.3,<2
joblib>=1.3,<2
mlflow>=2.9,<3
kafka-python>=2.1,<3
torch>=2.2,<3
